Refactoring of metrics
