+++ docs/command/mongocli-atlas-metrics-disks-describe.txt
@@ -19,7 +19,7 @@ Syntax
 
 .. code-block::
 
-   mongocli atlas metrics disks describe <hostname:port> <name> [options]
+   mongocli atlas metrics disks describe <hostname:port> <diskName> [options]
 
 Options
 -------
@@ -88,4 +88,12 @@ Inherited Options
      - string
      - false
      - Profile to use from your configuration file.
+
+Examples
+--------
+
+.. code-block::
+
+
+   $ mongocli atlas metrics disk describe atlas-lnmtkm-shard-00-00.ajlj3.mongodb.net:27017 Cluster0 --granularity PT1M --period P1DT12H
