+++ internal/cli/atlas/metrics/databases/databases.go
@@ -12,24 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package metrics
+package databases
 
 import (
 	"github.com/mongodb/mongocli/internal/cli"
 	"github.com/spf13/cobra"
 )
 
-func DisksBuilder() *cobra.Command {
-	const use = "disks"
+func Builder() *cobra.Command {
+	const use = "databases"
 	cmd := &cobra.Command{
 		Use:     use,
 		Aliases: cli.GenerateAliases(use),
-		Short:   "List available disks or disks measurements for a given host.",
+		Short:   "List available databases or databases measurements for a given host.",
 	}
-
 	cmd.AddCommand(
-		DisksListBuilder(),
-		DisksDescribeBuilder(),
+		ListBuilder(),
+		DescribeBuilder(),
 	)
 
 	return cmd
