+++ internal/cli/atlas/metrics/databases/describe.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package metrics
+package databases
 
 import (
 	"context"
@@ -26,16 +26,17 @@ import (
 	"github.com/spf13/cobra"
 )
 
-type ProcessOpts struct {
+type DescribeOpts struct {
 	cli.GlobalOpts
 	cli.OutputOpts
 	cli.MetricsOpts
 	host  string
 	port  int
-	store store.ProcessMeasurementLister
+	name  string
+	store store.ProcessDatabaseMeasurementsLister
 }
 
-func (opts *ProcessOpts) initStore(ctx context.Context) func() error {
+func (opts *DescribeOpts) initStore(ctx context.Context) func() error {
 	return func() error {
 		var err error
 		opts.store, err = store.New(store.AuthenticatedPreset(config.Default()), store.WithContext(ctx))
@@ -43,9 +44,14 @@ func (opts *ProcessOpts) initStore(ctx context.Context) func() error {
 	}
 }
 
-func (opts *ProcessOpts) Run() error {
+var databasesMetricTemplate = `NAME	UNITS	TIMESTAMP		VALUE{{range .ProcessMeasurements.Measurements}}  {{if .DataPoints}}
+{{- $name := .Name }}{{- $unit := .Units }}{{- range .DataPoints}}	
+{{ $name }}	{{ $unit }}	{{.Timestamp}}	{{if .Value }}	{{ .Value }}{{else}}	N/A {{end}}{{end}}{{end}}{{end}}
+`
+
+func (opts *DescribeOpts) Run() error {
 	listOpts := opts.NewProcessMetricsListOptions()
-	r, err := opts.store.ProcessMeasurements(opts.ConfigProjectID(), opts.host, opts.port, listOpts)
+	r, err := opts.store.ProcessDatabaseMeasurements(opts.ConfigProjectID(), opts.host, opts.port, opts.name, listOpts)
 	if err != nil {
 		return err
 	}
@@ -53,39 +59,46 @@ func (opts *ProcessOpts) Run() error {
 	return opts.Print(r)
 }
 
-var metricTemplate = `NAME	UNITS	TIMESTAMP		VALUE{{range .Measurements}} {{if .DataPoints}}
-{{- $name := .Name }}{{- $unit := .Units }}{{- range .DataPoints}}	
-{{ $name }}	{{ $unit }}	{{.Timestamp}}	{{if .Value }}	{{ .Value }}{{else}}	N/A {{end}}{{end}}{{end}}{{end}}
-`
-
-// mongocli atlas metric(s) process(es) <hostname:port> [--granularity granularity] [--period period] [--start start] [--end end] [--type type][--projectId projectId].
-func ProcessBuilder() *cobra.Command {
-	opts := &ProcessOpts{}
+// mcli atlas metric(s) database(s) describe <host:port> <databaseName> --granularity g --period p --start start --end end [--type type] [--projectId projectId].
+func DescribeBuilder() *cobra.Command {
+	const argsN = 2
+	opts := &DescribeOpts{}
 	cmd := &cobra.Command{
-		Use:     "processes <hostname:port>",
-		Short:   "Get measurements for a given host.",
-		Aliases: []string{"process"},
-		Args:    require.ExactArgs(1),
+		Use:   "describe <host:port> <databaseName>",
+		Short: "Describe database measurements for a given host database.",
+		Args:  require.ExactArgs(argsN),
+		Annotations: map[string]string{
+			"arg1":              "hostname:port",
+			"requiredArg1":      "hostname:port",
+			"hostname:portDesc": "Hostname and port number of the machine running the Atlas MongoDB process.",
+			"arg2":              "databaseName",
+			"requiredArg2":      "databaseName",
+			"databaseNameDesc":  "Label that identifies the database from which you want to retrieve measurements.",
+		},
+		Example: `
+  $ mongocli atlas metrics database describe atlas-lnmtkm-shard-00-00.ajlj3.mongodb.net:27017 Cluster0 --granularity PT1M --period P1DT12H`,
 		PreRunE: func(cmd *cobra.Command, args []string) error {
 			return opts.PreRunE(
 				opts.ValidateProjectID,
 				opts.ValidatePeriodStartEnd,
 				opts.initStore(cmd.Context()),
-				opts.InitOutput(cmd.OutOrStdout(), metricTemplate),
+				opts.InitOutput(cmd.OutOrStdout(), databasesMetricTemplate),
 			)
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
 			var err error
-			opts.host, opts.port, err = getHostnameAndPort(args[0])
+			opts.host, opts.port, err = opts.GetHostnameAndPort(args[0])
 			if err != nil {
 				return err
 			}
+			opts.name = args[1]
 
 			return opts.Run()
 		},
 	}
-	cmd.Flags().IntVar(&opts.PageNum, flag.Page, cli.DefaultPage, usage.Page)
-	cmd.Flags().IntVar(&opts.ItemsPerPage, flag.Limit, cli.DefaultPageLimit, usage.Limit)
+
+	cmd.Flags().IntVar(&opts.PageNum, flag.Page, 0, usage.Page)
+	cmd.Flags().IntVar(&opts.ItemsPerPage, flag.Limit, 0, usage.Limit)
 
 	cmd.Flags().StringVar(&opts.Granularity, flag.Granularity, "", usage.Granularity)
 	cmd.Flags().StringVar(&opts.Period, flag.Period, "", usage.Period)
