+++ internal/cli/atlas/metrics/databases/list.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package metrics
+package databases
 
 import (
 	"context"
@@ -26,16 +26,17 @@ import (
 	"github.com/spf13/cobra"
 )
 
-type DatabasesListsOpts struct {
+type ListsOpts struct {
 	cli.GlobalOpts
 	cli.OutputOpts
+	cli.MetricsOpts
 	cli.ListOpts
 	host  string
 	port  int
 	store store.ProcessDatabaseLister
 }
 
-func (opts *DatabasesListsOpts) initStore(ctx context.Context) func() error {
+func (opts *ListsOpts) initStore(ctx context.Context) func() error {
 	return func() error {
 		var err error
 		opts.store, err = store.New(store.AuthenticatedPreset(config.Default()), store.WithContext(ctx))
@@ -47,7 +48,7 @@ var databasesListTemplate = `{{range .Results}}
 {{.DatabaseName}}{{end}}
 `
 
-func (opts *DatabasesListsOpts) Run() error {
+func (opts *ListsOpts) Run() error {
 	listOpts := opts.NewListOptions()
 	r, err := opts.store.ProcessDatabases(opts.ConfigProjectID(), opts.host, opts.port, listOpts)
 	if err != nil {
@@ -57,14 +58,21 @@ func (opts *DatabasesListsOpts) Run() error {
 	return opts.Print(r)
 }
 
-// mongocli atlas metric(s) process(es) disks lists <hostname:port>.
-func DatabasesListBuilder() *cobra.Command {
-	opts := &DatabasesListsOpts{}
+// mongocli atlas metric(s) database(s) lists <hostname:port>.
+func ListBuilder() *cobra.Command {
+	opts := &ListsOpts{}
 	cmd := &cobra.Command{
 		Use:     "list <hostname:port>",
 		Short:   "List available databases for a given host.",
 		Aliases: []string{"ls"},
-		Args:    require.ExactArgs(1),
+		Annotations: map[string]string{
+			"args":              "hostname:port",
+			"requiredArgs":      "hostname:port",
+			"hostname:portDesc": "Hostname and port number of the machine running the Atlas MongoDB process.",
+		},
+		Example: `
+  $ mongocli atlas metrics database ls atlas-lnmtkm-shard-00-00.ajlj3.mongodb.net:27017`,
+		Args: require.ExactArgs(1),
 		PreRunE: func(cmd *cobra.Command, args []string) error {
 			return opts.PreRunE(
 				opts.ValidateProjectID,
@@ -74,7 +82,7 @@ func DatabasesListBuilder() *cobra.Command {
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
 			var err error
-			if opts.host, opts.port, err = getHostnameAndPort(args[0]); err != nil {
+			if opts.host, opts.port, err = opts.GetHostnameAndPort(args[0]); err != nil {
 				return err
 			}
