+++ internal/cli/atlas/metrics/disks/describe_test.go
@@ -15,22 +15,27 @@
 //go:build unit
 // +build unit
 
-package metrics
+package disks
 
 import (
 	"testing"
+
+	"github.com/mongodb/mongocli/internal/flag"
+	"github.com/mongodb/mongocli/internal/test"
 
 	"github.com/golang/mock/gomock"
 	"github.com/mongodb/mongocli/internal/mocks"
 	"go.mongodb.org/atlas/mongodbatlas"
 )
 
-func TestDatabasesDescribeOpts_Run(t *testing.T) {
+func TestDisksDescribeOpts_Run(t *testing.T) {
 	ctrl := gomock.NewController(t)
-	mockStore := mocks.NewMockProcessDatabaseMeasurementsLister(ctrl)
+	mockStore := mocks.NewMockProcessDiskMeasurementsLister(ctrl)
 	defer ctrl.Finish()
 
-	listOpts := &DatabasesDescribeOpts{
+	expected := &mongodbatlas.ProcessDiskMeasurements{}
+
+	listOpts := &DescribeOpts{
 		host:  "hard-00-00.mongodb.net",
 		port:  27017,
 		name:  "test",
@@ -38,17 +43,23 @@ func TestDatabasesDescribeOpts_Run(t *testing.T) {
 	}
 
 	opts := listOpts.NewProcessMetricsListOptions()
-	expected := &mongodbatlas.ProcessDatabaseMeasurements{
-		ProcessMeasurements: &mongodbatlas.ProcessMeasurements{
-			Measurements: []*mongodbatlas.Measurements{},
-		},
-	}
 	mockStore.
-		EXPECT().ProcessDatabaseMeasurements(listOpts.ProjectID, listOpts.host, listOpts.port, listOpts.name, opts).
+		EXPECT().ProcessDiskMeasurements(listOpts.ProjectID, listOpts.host, listOpts.port, listOpts.name, opts).
 		Return(expected, nil).
 		Times(1)
 
 	if err := listOpts.Run(); err != nil {
 		t.Fatalf("Run() unexpected error: %v", err)
 	}
+}
+
+func TestDescribeBuilder(t *testing.T) {
+	test.CmdValidator(
+		t,
+		DescribeBuilder(),
+		0,
+		[]string{
+			flag.Page, flag.Limit, flag.Granularity, flag.Period, flag.Start,
+			flag.End, flag.Type, flag.ProjectID, flag.Output},
+	)
 }
