+++ internal/cli/atlas/metrics/disks/list_test.go
@@ -11,17 +11,18 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
 //go:build unit
 // +build unit
 
-package metrics
+package disks
 
 import (
 	"testing"
 
 	"github.com/golang/mock/gomock"
+	"github.com/mongodb/mongocli/internal/flag"
 	"github.com/mongodb/mongocli/internal/mocks"
+	"github.com/mongodb/mongocli/internal/test"
 	"go.mongodb.org/atlas/mongodbatlas"
 )
 
@@ -32,7 +33,7 @@ func TestDisksListsOpts_Run(t *testing.T) {
 
 	expected := &mongodbatlas.ProcessDisksResponse{}
 
-	listOpts := &DisksListsOpts{
+	listOpts := &ListsOpts{
 		host:  "hard-00-00.mongodb.net",
 		port:  27017,
 		store: mockStore,
@@ -47,4 +48,13 @@ func TestDisksListsOpts_Run(t *testing.T) {
 	if err := listOpts.Run(); err != nil {
 		t.Fatalf("Run() unexpected error: %v", err)
 	}
+}
+
+func TestListBuilder(t *testing.T) {
+	test.CmdValidator(
+		t,
+		ListBuilder(),
+		0,
+		[]string{flag.Page, flag.Limit, flag.ProjectID, flag.Output},
+	)
 }
