+++ internal/cli/atlas/metrics/processes/processes.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package metrics
+package processes
 
 import (
 	"context"
@@ -26,16 +26,16 @@ import (
 	"github.com/spf13/cobra"
 )
 
-type DisksListsOpts struct {
+type Opts struct {
 	cli.GlobalOpts
 	cli.OutputOpts
-	cli.ListOpts
+	cli.MetricsOpts
 	host  string
 	port  int
-	store store.ProcessDisksLister
+	store store.ProcessMeasurementLister
 }
 
-func (opts *DisksListsOpts) initStore(ctx context.Context) func() error {
+func (opts *Opts) initStore(ctx context.Context) func() error {
 	return func() error {
 		var err error
 		opts.store, err = store.New(store.AuthenticatedPreset(config.Default()), store.WithContext(ctx))
@@ -43,9 +43,9 @@ func (opts *DisksListsOpts) initStore(ctx context.Context) func() error {
 	}
 }
 
-func (opts *DisksListsOpts) Run() error {
-	listOpts := opts.NewListOptions()
-	r, err := opts.store.ProcessDisks(opts.ConfigProjectID(), opts.host, opts.port, listOpts)
+func (opts *Opts) Run() error {
+	listOpts := opts.NewProcessMetricsListOptions()
+	r, err := opts.store.ProcessMeasurements(opts.ConfigProjectID(), opts.host, opts.port, listOpts)
 	if err != nil {
 		return err
 	}
@@ -53,28 +53,37 @@ func (opts *DisksListsOpts) Run() error {
 	return opts.Print(r)
 }
 
-var listTemplate = `{{range .Results}}
-{{.PartitionName}}{{end}}
+var metricTemplate = `NAME	UNITS	TIMESTAMP		VALUE{{range .Measurements}} {{if .DataPoints}}
+{{- $name := .Name }}{{- $unit := .Units }}{{- range .DataPoints}}	
+{{ $name }}	{{ $unit }}	{{.Timestamp}}	{{if .Value }}	{{ .Value }}{{else}}	N/A {{end}}{{end}}{{end}}{{end}}
 `
 
-// mongocli atlas metric(s) process(es) disks lists <hostname:port>.
-func DisksListBuilder() *cobra.Command {
-	opts := &DisksListsOpts{}
+// mongocli atlas metric(s) process(es) <hostname:port> [--granularity granularity] [--period period] [--start start] [--end end] [--type type][--projectId projectId].
+func Builder() *cobra.Command {
+	opts := &Opts{}
 	cmd := &cobra.Command{
-		Use:     "list <hostname:port>",
-		Short:   "List available disks for a given host.",
-		Aliases: []string{"ls"},
+		Use:     "processes <hostname:port>",
+		Short:   "Get measurements for a given host.",
+		Aliases: []string{"process"},
 		Args:    require.ExactArgs(1),
+		Annotations: map[string]string{
+			"args":              "hostname:port",
+			"requiredArgs":      "hostname:port",
+			"hostname:portDesc": "Hostname and port number of the machine running the Atlas MongoDB process.",
+		},
+		Example: `
+  $ mongocli atlas metrics process atlas-lnmtkm-shard-00-00.ajlj3.mongodb.net:27017`,
 		PreRunE: func(cmd *cobra.Command, args []string) error {
 			return opts.PreRunE(
 				opts.ValidateProjectID,
+				opts.ValidatePeriodStartEnd,
 				opts.initStore(cmd.Context()),
-				opts.InitOutput(cmd.OutOrStdout(), listTemplate),
+				opts.InitOutput(cmd.OutOrStdout(), metricTemplate),
 			)
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
 			var err error
-			opts.host, opts.port, err = getHostnameAndPort(args[0])
+			opts.host, opts.port, err = opts.GetHostnameAndPort(args[0])
 			if err != nil {
 				return err
 			}
@@ -82,11 +91,19 @@ func DisksListBuilder() *cobra.Command {
 			return opts.Run()
 		},
 	}
-
 	cmd.Flags().IntVar(&opts.PageNum, flag.Page, cli.DefaultPage, usage.Page)
 	cmd.Flags().IntVar(&opts.ItemsPerPage, flag.Limit, cli.DefaultPageLimit, usage.Limit)
+
+	cmd.Flags().StringVar(&opts.Granularity, flag.Granularity, "", usage.Granularity)
+	cmd.Flags().StringVar(&opts.Period, flag.Period, "", usage.Period)
+	cmd.Flags().StringVar(&opts.Start, flag.Start, "", usage.MeasurementStart)
+	cmd.Flags().StringVar(&opts.End, flag.End, "", usage.MeasurementEnd)
+	cmd.Flags().StringSliceVar(&opts.MeasurementType, flag.Type, nil, usage.MeasurementType)
+
 	cmd.Flags().StringVar(&opts.ProjectID, flag.ProjectID, "", usage.ProjectID)
 	cmd.Flags().StringVarP(&opts.Output, flag.Output, flag.OutputShort, "", usage.FormatOut)
+
+	_ = cmd.MarkFlagRequired(flag.Granularity)
 
 	return cmd
 }
