+++ internal/cli/metrics_opts_test.go
@@ -89,3 +89,33 @@ func TestMetricsOpts_ValidatePeriodStartEnd(t *testing.T) {
 		})
 	}
 }
+
+func TestGetHostNameAndPort(t *testing.T) {
+	t.Run("valid parameter", func(t *testing.T) {
+		opts := &MetricsOpts{}
+		host, port, err := opts.GetHostnameAndPort("test:2000")
+		if err != nil {
+			t.Fatalf("getHostnameAndPort unexpecteted err: %#v\n", err)
+		}
+		if host != "test" {
+			t.Errorf("Expected '%s', got '%s'\n", "test", host)
+		}
+		if port != 2000 {
+			t.Errorf("Expected '%d', got '%d'\n", 2000, port)
+		}
+	})
+	t.Run("incomplete format", func(t *testing.T) {
+		opts := &MetricsOpts{}
+		_, _, err := opts.GetHostnameAndPort("test")
+		if err == nil {
+			t.Fatal("getHostnameAndPort should return an error\n")
+		}
+	})
+	t.Run("incomplete format", func(t *testing.T) {
+		opts := &MetricsOpts{}
+		_, _, err := opts.GetHostnameAndPort(":test")
+		if err == nil {
+			t.Fatal("getHostnameAndPort should return an error\n")
+		}
+	})
+}
